fix iterdataset max_steps

+++ nlp/utils/__init__.py
@@ -19,7 +19,7 @@ def configure_optimizers(named_parameter: typing.Union[typing.List,typing.Tuple]
 
     if estimated_stepping_batches is None:
         if training_args.max_steps is not None and training_args.max_steps > 0:
-            estimated_stepping_batches = training_args.max_steps / training_args.gradient_accumulation_steps
+            estimated_stepping_batches = int(training_args.max_steps / training_args.gradient_accumulation_steps)
     assert estimated_stepping_batches is not None
 
     optimizer_name = training_args.optimizer.lower()
